Removed spacebars. Added Premium to reduce=. Etc.
Removed blank spaces.
Changed |reduce= to |reduce=premium as this only runs on gacha page.
Added {{Weapon/Ougi| ** }} to |ougi=, unsure if that is doable because of the slash. If it isn't, comment it out or remove?
Removed |evo_red= as it's never appeared in gacha.
Blanked |evo_max= in case it is a 4* item. It'll default to 3 regardless and I'll fill it in on the manual pass.